Remove store view, remove delivery time mageplaza
