+++ src/simi/App/nativeInner/Checkout/DeliveryDateTime/deliveryDateTime.gql.js
@@ -1,29 +1,31 @@
 import gql from 'graphql-tag';
 
-const GET_DELIVERY_TIME = gql`
-    query DeliveryInformation {
-        deliveryTime {
-            deliveryDateFormat
-            deliveryTime
-            deliveryDateOff
-            isEnabledDeliveryTime
-            deliveryDaysOff
-            isEnabledHouseSecurityCode
-            isEnabledDeliveryComment
+const DELIVERY_TIME_MUTATION = gql`
+    mutation(
+        $cart_id: String!
+        $shipping_arrival_comments: String
+        $shipping_arrival_date: String
+        $shipping_arrival_timeslot: String
+    ) {
+        applyDeliveryCart(
+            input: {
+                cart_id: $cart_id
+                shipping_arrival_comments: $shipping_arrival_comments
+                shipping_arrival_timeslot: $shipping_arrival_timeslot
+                shipping_arrival_date: $shipping_arrival_date
+            }
+        ) {
+            cart {
+                error_message
+                shipping_arrival_comments
+                shipping_arrival_date
+                shipping_arrival_timeslot
+                success
+            }
         }
     }
 `;
 
-const DELIVERY_TIME_MUTATION = gql`
-    mutation DeliveryMutation(
-        $cart_id: String!
-        $mp_delivery_time: DeliveryTimeInput!
-    ) {
-        MpDeliveryTime(cart_id: $cart_id, mp_delivery_time: $mp_delivery_time)
-    }
-`;
-
 export default {
-    getDeliveryTime: GET_DELIVERY_TIME,
     deliveryTimeMutation: DELIVERY_TIME_MUTATION
 };
