--- src/simi/App/nativeInner/Checkout/DeliveryDateTimeNew/index.js
@@ -1,1 +1,0 @@
-export { default } from './deliveryDateTime';
